Add files via upload
This is the server authentication code for registration and login details.

The server.php file takes in username, email and password and checks the database for match then throws an error if there is a match or record the details if there isn't. 

The errors.php throws an error alert for password mismatch and registration errors.